Correct canonical Pmetrics CSV compatibility
