+++ Cargo.toml
@@ -73,6 +73,7 @@ quick_cache = "0.7.0"
 wasmtime = { version = "44.0.0", optional = true }
 
 [dev-dependencies]
+bincode = "1.3.3"
 criterion = { version = "0.8.2", features = ["html_reports"] }
 approx = "0.5.1"
 tempfile = "3.27.0"
