+++ src/data/parser/pmetrics.rs
@@ -8,18 +8,18 @@
 //! numeric values such as `1` are preserved as numeric-looking labels.
 
 use crate::{data::*, PharmsolError};
-use csv::{ReaderBuilder, StringRecord, Terminator, WriterBuilder};
+use csv::{ReaderBuilder, StringRecord, WriterBuilder};
 use serde::de::{MapAccess, Visitor};
 use serde::{de, Deserialize, Deserializer, Serialize};
-use std::collections::{BTreeMap, HashMap, HashSet};
+use std::collections::{HashMap, HashSet};
 use std::fmt;
 use std::fs::File;
-use std::io::{Cursor, Read, Write};
+use std::io::{Cursor, Read};
 use std::str::FromStr;
 
 use crate::data::row::{build_data, DataError, DataRow};
 
-const CORE_HEADERS: [&str; 15] = [
+pub(super) const CORE_HEADERS: [&str; 15] = [
     "ID", "EVID", "TIME", "DUR", "DOSE", "ADDL", "II", "INPUT", "OUT", "OUTEQ", "CENS", "C0", "C1",
     "C2", "C3",
 ];
@@ -108,7 +108,14 @@ pub fn read_pmetrics_csv_v1<R: Read>(mut reader: R) -> Result<Data, DataError> {
         ));
     }
 
-    read_pmetrics_reader(Cursor::new(bytes), true)
+    let data = read_pmetrics_reader(Cursor::new(bytes.as_slice()), true)?;
+    let canonical = data.to_pmetrics_csv_v1()?;
+    if canonical != bytes {
+        return Err(DataError::InvalidCanonicalFormat(
+            "input bytes are not canonical pmetrics-csv.v1".to_string(),
+        ));
+    }
+    Ok(data)
 }
 
 fn read_pmetrics_reader<R: Read>(reader: R, canonical: bool) -> Result<Data, DataError> {
@@ -216,7 +223,7 @@ fn validate_canonical_headers(headers: &StringRecord) -> Result<(), DataError> {
     Ok(())
 }
 
-fn validate_covariate_header(header: &str) -> Result<(), DataError> {
+pub(super) fn validate_covariate_header(header: &str) -> Result<(), DataError> {
     let base = header.strip_suffix('!').unwrap_or(header);
     if base.is_empty()
         || base.ends_with('!')
@@ -230,6 +237,17 @@ fn validate_covariate_header(header: &str) -> Result<(), DataError> {
         )));
     }
     Ok(())
+}
+
+fn ensure_finite(value: f64, field: &str, id: &str) -> Result<(), DataError> {
+    if value.is_finite() {
+        Ok(())
+    } else {
+        Err(DataError::NonFiniteValue {
+            field: field.to_string(),
+            id: id.to_string(),
+        })
+    }
 }
 
 /// One row from a Pmetrics file after serde deserialization.
@@ -597,327 +615,124 @@ where
     deserializer.deserialize_map(CovsVisitor)
 }
 
-#[derive(Debug)]
-struct CanonicalCovariate {
-    name: String,
-    header: String,
-}
-
-#[derive(Debug)]
-struct CanonicalRow {
-    time: f64,
-    rank: u8,
-    sequence: usize,
-    fields: Vec<String>,
-}
-
-fn ensure_finite(value: f64, field: &str, id: &str) -> Result<(), DataError> {
-    if value.is_finite() {
-        Ok(())
-    } else {
-        Err(DataError::NonFiniteValue {
-            field: field.to_string(),
-            id: id.to_string(),
-        })
-    }
-}
-
-fn ensure_label(label: &str, field: &str, id: &str) -> Result<(), DataError> {
-    if label.is_empty() || label == "." || label == "NA" || label.contains('\r') {
-        Err(DataError::InvalidCanonicalFormat(format!(
-            "{field} label `{label}` for {id} is reserved as missing"
-        )))
-    } else {
-        Ok(())
-    }
-}
-
-fn collect_covariate_schema(data: &Data) -> Result<Vec<CanonicalCovariate>, DataError> {
-    let mut fixed_by_name = BTreeMap::<String, bool>::new();
-    let mut folded_names = HashMap::<String, String>::new();
-
-    for subject in data.subjects() {
-        for occasion in subject.occasions() {
-            for (key, covariate) in occasion.covariates().covariates() {
-                if key != covariate.name() {
-                    return Err(DataError::InvalidCanonicalFormat(format!(
-                        "covariate key `{key}` does not match name `{}`",
-                        covariate.name()
-                    )));
-                }
-                validate_covariate_header(&key)?;
-                if key.ends_with('!') {
-                    return Err(DataError::InvalidCanonicalFormat(format!(
-                        "covariate name `{key}` reserves trailing ! for fixed covariates"
-                    )));
-                }
-                let folded = key.to_lowercase();
-                if let Some(existing) = folded_names.insert(folded, key.clone()) {
-                    if existing != key {
-                        return Err(DataError::InvalidCanonicalFormat(format!(
-                            "covariate names `{existing}` and `{key}` are ambiguous"
-                        )));
-                    }
-                }
-                if covariate.observations().is_empty() {
-                    return Err(DataError::InvalidCanonicalFormat(format!(
-                        "covariate `{key}` for subject `{}` occasion {} has no observations",
-                        subject.id(),
-                        occasion.index()
-                    )));
-                }
-                if let Some(existing) = fixed_by_name.insert(key.clone(), covariate.fixed()) {
-                    if existing != covariate.fixed() {
-                        return Err(DataError::InvalidCanonicalFormat(format!(
-                            "covariate `{key}` has inconsistent fixed semantics"
-                        )));
-                    }
-                }
-            }
-        }
-    }
-
-    let mut schema = fixed_by_name
-        .into_iter()
-        .map(|(name, fixed)| CanonicalCovariate {
-            header: if fixed {
-                format!("{name}!")
-            } else {
-                name.clone()
-            },
-            name,
-        })
-        .collect::<Vec<_>>();
-    schema.sort_by(|left, right| left.header.cmp(&right.header));
-    Ok(schema)
-}
-
-fn empty_row(id: &str, evid: i32, time: f64, covariate_count: usize) -> Vec<String> {
-    let mut fields = vec![".".to_string(); CORE_HEADERS.len() + covariate_count];
-    fields[0] = id.to_string();
-    fields[1] = evid.to_string();
-    fields[2] = time.to_string();
-    fields
-}
-
-fn event_rank(event: &Event) -> u8 {
-    match event {
-        Event::Observation(_) => 1,
-        Event::Bolus(_) => 2,
-        Event::Infusion(_) => 3,
-    }
-}
-
-fn event_row(id: &str, event: &Event, covariate_count: usize) -> Result<Vec<String>, DataError> {
-    ensure_finite(event.time(), "TIME", id)?;
-    let mut fields = empty_row(
-        id,
-        match event {
-            Event::Observation(_) => 0,
-            Event::Bolus(_) | Event::Infusion(_) => 1,
-        },
-        event.time(),
-        covariate_count,
-    );
-
-    match event {
-        Event::Observation(observation) => {
-            if let Some(value) = observation.value() {
-                ensure_finite(value, "OUT", id)?;
-                fields[8] = value.to_string();
-            }
-            let outeq = observation.outeq().to_string();
-            ensure_label(&outeq, "OUTEQ", id)?;
-            fields[9] = outeq;
-            fields[10] = match observation.censoring() {
-                Censor::None => "0",
-                Censor::BLOQ => "1",
-                Censor::ALOQ => "-1",
-            }
-            .to_string();
-            if let Some(error) = observation.errorpoly() {
-                let coefficients = error.coefficients();
-                for (offset, value) in [
-                    coefficients.0,
-                    coefficients.1,
-                    coefficients.2,
-                    coefficients.3,
-                ]
-                .into_iter()
-                .enumerate()
-                {
-                    ensure_finite(value, &format!("C{offset}"), id)?;
-                    fields[11 + offset] = value.to_string();
-                }
-            }
-        }
-        Event::Bolus(bolus) => {
-            ensure_finite(bolus.amount(), "DOSE", id)?;
-            let input = bolus.input().to_string();
-            ensure_label(&input, "INPUT", id)?;
-            fields[3] = "0".to_string();
-            fields[4] = bolus.amount().to_string();
-            fields[7] = input;
-        }
-        Event::Infusion(infusion) => {
-            ensure_finite(infusion.duration(), "DUR", id)?;
-            ensure_finite(infusion.amount(), "DOSE", id)?;
-            if infusion.duration() <= 0.0 {
-                return Err(DataError::InvalidCanonicalFormat(format!(
-                    "infusion duration for {id} must be greater than zero"
-                )));
-            }
-            let input = infusion.input().to_string();
-            ensure_label(&input, "INPUT", id)?;
-            fields[3] = infusion.duration().to_string();
-            fields[4] = infusion.amount().to_string();
-            fields[7] = input;
-        }
-    }
-    Ok(fields)
-}
-
 impl Data {
-    /// Write canonical `pmetrics-csv.v1` bytes to a stream.
-    pub fn write_pmetrics_csv_v1<W: Write>(&self, writer: W) -> Result<(), DataError> {
-        let schema = collect_covariate_schema(self)?;
-        let mut headers = CORE_HEADERS
-            .iter()
-            .map(ToString::to_string)
-            .collect::<Vec<_>>();
-        headers.extend(schema.iter().map(|covariate| covariate.header.clone()));
-
-        let mut csv = WriterBuilder::new()
-            .has_headers(false)
-            .terminator(Terminator::Any(b'\n'))
-            .from_writer(writer);
-        csv.write_record(&headers)
-            .map_err(|error| DataError::CSVError(error.to_string()))?;
-
-        let mut subjects = self.subjects();
-        subjects.sort_by(|left, right| left.id().cmp(right.id()));
-        for pair in subjects.windows(2) {
-            if pair[0].id() == pair[1].id() {
-                return Err(DataError::InvalidCanonicalFormat(format!(
-                    "duplicate subject ID `{}`",
-                    pair[0].id()
-                )));
-            }
-        }
-
-        for subject in subjects {
-            if subject.id().contains('\r') {
-                return Err(DataError::InvalidCanonicalFormat(format!(
-                    "subject ID `{}` contains a carriage return",
-                    subject.id()
-                )));
-            }
-            if subject.occasions().is_empty() {
-                return Err(DataError::InvalidCanonicalFormat(format!(
-                    "subject `{}` has no occasions",
-                    subject.id()
-                )));
-            }
-            for (occasion_index, occasion) in subject.occasions().iter().enumerate() {
-                if occasion.index() != occasion_index {
-                    return Err(DataError::InvalidCanonicalFormat(format!(
-                        "subject `{}` has nonsequential occasion index {}",
-                        subject.id(),
-                        occasion.index()
-                    )));
-                }
-
-                let events = occasion.events();
-                for event in events {
-                    if event.occasion() != occasion_index {
-                        return Err(DataError::InvalidCanonicalFormat(format!(
-                            "subject `{}` has an event assigned to occasion {} inside occasion {}",
-                            subject.id(),
-                            event.occasion(),
-                            occasion_index
-                        )));
-                    }
-                }
-                for pair in events.windows(2) {
-                    let order = pair[0]
-                        .time()
-                        .total_cmp(&pair[1].time())
-                        .then_with(|| event_rank(&pair[0]).cmp(&event_rank(&pair[1])));
-                    if order.is_gt() {
-                        return Err(DataError::InvalidCanonicalFormat(format!(
-                            "events for subject `{}` occasion {} are not in canonical order",
-                            subject.id(),
-                            occasion_index
-                        )));
-                    }
-                }
-
-                let mut rows = Vec::new();
-                for (sequence, event) in events.iter().enumerate() {
-                    rows.push(CanonicalRow {
-                        time: event.time(),
-                        rank: event_rank(event),
-                        sequence,
-                        fields: event_row(subject.id(), event, schema.len())?,
-                    });
-                }
-
-                let covariates = occasion.covariates().covariates();
-                let mut covariate_rows = HashMap::<u64, CanonicalRow>::new();
-                for (column, canonical) in schema.iter().enumerate() {
-                    let Some(covariate) = covariates.get(&canonical.name) else {
-                        continue;
-                    };
-                    for (time, value) in covariate.observations() {
-                        ensure_finite(time, &format!("{} time", canonical.name), subject.id())?;
-                        ensure_finite(value, &canonical.name, subject.id())?;
-                        let next_sequence = rows.len() + covariate_rows.len();
-                        let row =
-                            covariate_rows
-                                .entry(time.to_bits())
-                                .or_insert_with(|| CanonicalRow {
-                                    time,
-                                    rank: 4,
-                                    sequence: next_sequence,
-                                    fields: empty_row(subject.id(), 2, time, schema.len()),
-                                });
-                        row.fields[CORE_HEADERS.len() + column] = value.to_string();
-                    }
-                }
-                rows.extend(covariate_rows.into_values());
-                rows.sort_by(|left, right| {
-                    left.time
-                        .total_cmp(&right.time)
-                        .then_with(|| left.rank.cmp(&right.rank))
-                        .then_with(|| left.sequence.cmp(&right.sequence))
-                });
-
-                let boundary_time = rows.first().map_or(0.0, |row| row.time);
-                ensure_finite(boundary_time, "TIME", subject.id())?;
-                csv.write_record(empty_row(subject.id(), 4, boundary_time, schema.len()))
-                    .map_err(|error| DataError::CSVError(error.to_string()))?;
-                for row in rows {
-                    csv.write_record(row.fields)
-                        .map_err(|error| DataError::CSVError(error.to_string()))?;
-                }
-            }
-        }
-
-        csv.flush()
-            .map_err(|error| DataError::CSVError(error.to_string()))
-    }
-
-    /// Return canonical `pmetrics-csv.v1` bytes.
-    pub fn to_pmetrics_csv_v1(&self) -> Result<Vec<u8>, DataError> {
-        let mut bytes = Vec::new();
-        self.write_pmetrics_csv_v1(&mut bytes)?;
-        Ok(bytes)
-    }
-
-    /// Write the dataset to a file in canonical Pmetrics format.
+    /// Write the dataset to a file in ordinary Pmetrics format.
+    ///
+    /// `INPUT` and `OUTEQ` are written using their stored public labels. Named
+    /// labels such as `iv` and `cp` remain named labels, and numeric-looking
+    /// labels are written back exactly as stored.
+    ///
+    /// Missing optional fields are emitted as `.` placeholders to match the
+    /// usual Pmetrics text convention.
     pub fn write_pmetrics(&self, file: &File) -> Result<(), PharmsolError> {
-        self.write_pmetrics_csv_v1(file)
-            .map_err(PharmsolError::from)
+        let mut writer = WriterBuilder::new().has_headers(true).from_writer(file);
+
+        writer
+            .write_record(CORE_HEADERS)
+            .map_err(|error| PharmsolError::OtherError(error.to_string()))?;
+
+        for subject in self.subjects() {
+            for occasion in subject.occasions() {
+                for event in occasion.process_events(None) {
+                    match event {
+                        Event::Observation(observation) => {
+                            let time = observation.time().to_string();
+                            let value = observation
+                                .value()
+                                .map_or_else(|| ".".to_string(), |value| value.to_string());
+                            let outeq = observation.outeq().to_string();
+                            let censor = match observation.censoring() {
+                                Censor::None => "0".to_string(),
+                                Censor::BLOQ => "1".to_string(),
+                                Censor::ALOQ => "-1".to_string(),
+                            };
+                            let (c0, c1, c2, c3) = observation
+                                .errorpoly()
+                                .map(|poly| {
+                                    let (c0, c1, c2, c3) = poly.coefficients();
+                                    (
+                                        c0.to_string(),
+                                        c1.to_string(),
+                                        c2.to_string(),
+                                        c3.to_string(),
+                                    )
+                                })
+                                .unwrap_or_else(|| {
+                                    (
+                                        ".".to_string(),
+                                        ".".to_string(),
+                                        ".".to_string(),
+                                        ".".to_string(),
+                                    )
+                                });
+
+                            writer
+                                .write_record([
+                                    subject.id(),
+                                    "0",
+                                    &time,
+                                    ".",
+                                    ".",
+                                    ".",
+                                    ".",
+                                    ".",
+                                    &value,
+                                    &outeq,
+                                    &censor,
+                                    &c0,
+                                    &c1,
+                                    &c2,
+                                    &c3,
+                                ])
+                                .map_err(|error| PharmsolError::OtherError(error.to_string()))?;
+                        }
+                        Event::Infusion(infusion) => {
+                            writer
+                                .write_record([
+                                    subject.id(),
+                                    "1",
+                                    &infusion.time().to_string(),
+                                    &infusion.duration().to_string(),
+                                    &infusion.amount().to_string(),
+                                    ".",
+                                    ".",
+                                    infusion.input().as_ref(),
+                                    ".",
+                                    ".",
+                                    ".",
+                                    ".",
+                                    ".",
+                                    ".",
+                                    ".",
+                                ])
+                                .map_err(|error| PharmsolError::OtherError(error.to_string()))?;
+                        }
+                        Event::Bolus(bolus) => {
+                            writer
+                                .write_record([
+                                    subject.id(),
+                                    "1",
+                                    &bolus.time().to_string(),
+                                    "0",
+                                    &bolus.amount().to_string(),
+                                    ".",
+                                    ".",
+                                    bolus.input().as_ref(),
+                                    ".",
+                                    ".",
+                                    ".",
+                                    ".",
+                                    ".",
+                                    ".",
+                                    ".",
+                                ])
+                                .map_err(|error| PharmsolError::OtherError(error.to_string()))?;
+                        }
+                    }
+                }
+            }
+        }
+        Ok(())
     }
 }
 
@@ -991,7 +806,7 @@ mod tests {
         let infusion_row = reader
             .records()
             .filter_map(Result::ok)
-            .find(|record| record.get(1) == Some("1") && record.get(3) != Some("0"))
+            .find(|record| record.get(3) != Some("0"))
             .expect("infusion row missing");
 
         assert_eq!(infusion_row.get(7), Some("3")); // Written as-is (1-indexed)
